Prepare for Causal RL 2.0 with GPT Integration and Gamma/Lambda Learning
- Implement integration of Causal RL algorithm with GPT model to facilitate advanced contextual action learning. This synergy aims to leverage the strengths of both models for improved decision-making capabilities.
- Introduce gamma and lambda as learnable network parameters, enabling the model to dynamically adjust its discounting and decision-making strategies based on the evolving context.
- Develop an automated control mechanism for GPT's input sequence, optimizing the model's ability to process and react to complex sequences of actions and outcomes.
- Launch a one-click RL solution, streamlining the deployment process for users. This feature is designed to lower the barrier to entry, making it easier for both beginners and experienced practitioners to leverage the power of Causal RL 2.0.

These enhancements are geared towards making Causal RL 2.0 a more robust and user-friendly platform for exploring and exploiting the potential of causal reasoning in reinforcement learning environments.

+++ utils/structure/env_observation.py
@@ -51,9 +51,9 @@ class EnvObservation:
     def shift(self, term_agents, dec_agents):
         # Convert term_agents and dec_agents to torch tensors if they are not already
         if not isinstance(term_agents, torch.Tensor):
-            term_agents = torch.tensor(term_agents, device=self.device)
+            term_agents = torch.tensor(term_agents, dtype=torch.long, device=self.device)
         if not isinstance(dec_agents, torch.Tensor):
-            dec_agents = torch.tensor(dec_agents, device=self.device)
+            dec_agents = torch.tensor(dec_agents, dtype=torch.long, device=self.device)
             
         # Roll data and mask for 'dec_agents'
         for key in self.data:
